Move code from commown to commown_payment_slimpay_issue
And move tests accordingly.

This makes the commown module lighter and eases code search.
