+++ commown/tests/test_project_task.py
@@ -262,37 +262,3 @@ class ProjectTaskActionTC(TransactionCase):
         self.assertEqual(
             self.task.stage_id, self.stage_pending, self.task.stage_id.name
         )
-
-    def test_payment_task_process_automatically(self):
-        inv = (
-            self.env["account.invoice"]
-            .search([])
-            .filtered(
-                lambda i: i.invoice_line_ids
-                and not any(line.contract_line_id for line in i.invoice_line_ids)
-            )[0]
-        )
-        self.task.invoice_id = inv.id
-        iline = inv.invoice_line_ids[0]
-
-        self.assertFalse(self.task.slimpay_payment_issue_process_automatically())
-
-        contract = self.env["contract.contract"].create(
-            {
-                "name": "Test Contract",
-                "partner_id": inv.partner_id.id,
-                "contract_line_ids": [
-                    (
-                        0,
-                        0,
-                        {
-                            "name": "line 1",
-                            "product_id": iline.product_id.id,
-                        },
-                    )
-                ],
-            }
-        )
-        iline.contract_line_id = contract.contract_line_ids[0].id
-
-        self.assertTrue(self.task.slimpay_payment_issue_process_automatically())
