+++ commown_payment_slimpay_issue/__init__.py
@@ -1,0 +1,1 @@
+from . import models
