+++ commown_payment_slimpay_issue/models/__init__.py
@@ -1,0 +1,1 @@
+from . import project_task
