+++ commown_payment_slimpay_issue/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_project_task
